Fixes for creating launchers
