+++ resources/constants.py
@@ -48,11 +48,12 @@ VLAUNCHER_MOST_PLAYED_ID = 'vlauncher_most_played'
 VCATEGORY_PCLONES_ID     = 'vcat_pclone'
 
 # LAUNCHER TYPES
-LAUNCHER_STANDALONE  = 1
-LAUNCHER_ROM         = 2
-LAUNCHER_RETROPLAYER = 3
-LAUNCHER_LNK         = 4
-LAUNCHER_RETROARCH   = 5
+LAUNCHER_STANDALONE  = 'STANDALONE'
+LAUNCHER_ROM         = 'ROM'
+LAUNCHER_RETROPLAYER = 'RETROPLAYER'
+LAUNCHER_LNK         = 'WINLNK'
+LAUNCHER_RETROARCH   = 'RETROARCH'
+LAUNCHER_STEAM       = 'STEAM'
 
 # --- Content type property to be used by skins ---
 AEL_CONTENT_WINDOW_ID       = 10000
